+++ src/index.ts
@@ -98,6 +98,22 @@ export {
   ObserveLocationChangeResponse,
 } from './modules/location';
 export {
+  EstimateRewardsEntryStatusCode,
+  EstimateRewardsErrorEntry,
+  EstimateRewardsErrorReasonCode,
+  EstimateRewardsEstimatedFiat,
+  EstimateRewardsNotApplicableEntry,
+  EstimateRewardsNotApplicableReasonCode,
+  EstimateRewardsRequest,
+  EstimateRewardsRequestEntry,
+  EstimateRewardsResponse,
+  EstimateRewardsResult,
+  EstimateRewardsResultEntry,
+  EstimateRewardsReward,
+  EstimateRewardsSuccessEntry,
+  LoyaltyModule,
+} from './modules/loyalty';
+export {
   DRMContentConfig,
   DRMPlaybackEvent,
   MediaModule,