+++ src/modules/loyalty/index.ts
@@ -1,0 +1,23 @@
+/*!
+ * Copyright (c) Grab Taxi Holdings PTE LTD (GRAB)
+ *
+ * This source code is licensed under the MIT license found in the LICENSE file in the root
+ * directory of this source tree.
+ */
+
+export { LoyaltyModule } from './LoyaltyModule';
+export type {
+  EstimateRewardsEntryStatusCode,
+  EstimateRewardsErrorEntry,
+  EstimateRewardsErrorReasonCode,
+  EstimateRewardsEstimatedFiat,
+  EstimateRewardsNotApplicableEntry,
+  EstimateRewardsNotApplicableReasonCode,
+  EstimateRewardsRequest,
+  EstimateRewardsRequestEntry,
+  EstimateRewardsResponse,
+  EstimateRewardsResult,
+  EstimateRewardsResultEntry,
+  EstimateRewardsReward,
+  EstimateRewardsSuccessEntry,
+} from './types';